pyproject.toml
[tool.poetry]
name = "starwars"
version = "0.1.0"
description = ""
authors = ["Erik Kvale <eirikval@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.65.1"
ariadne = "^0.13.0"
psycopg2-binary = "^2.8.6"
SQLAlchemy = "1.4.17"
python-dotenv = "^0.17.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
